Move IndirectPdfObjectReader again
