+++ Pdfer/ObjectRepository.cs
@@ -1,5 +1,4 @@
 using Pdfer.Objects;
-using Pdfer.Objects.ObjectReaders;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
