+++ Pdfer/Objects/IndirectPdfObjectReader.cs
